add edge cases

+++ src/lin_frame.h
@@ -15,7 +15,7 @@ typedef struct
     uint8_t checksum;
 } lin_frame_t;
 
-uint8_t lin_compute_pid(uint8_t id);
+int16_t lin_compute_pid(uint8_t id);
 uint8_t lin_compute_checksum(const lin_frame_t *frame, bool enhanced);
 
 #endif /*LIN_FRAME_H*/
